Add greedy quantifier regex mutator
In regex c{5,8} checks if character c matches 5 to 8 times consecutively. We can mutate this to c{4,8}, c{6,8}, c{5,7}, and c{5,9} to test edge cases. Invalid cases are skipped. E.g. when the end value becomes less than the start c{4,3} or the start becomes lower than 0 c{-1,3}.

Solves part of this issue:
https://github.com/stryker-mutator/stryker-net/issues/1228