feat(Regex mutator): Add mutations for regex greedy quantifier (#1276)
* Add greedy quantifier regex mutator

In regex c{5,8} checks if character c matches 5 to 8 times consecutively. We can mutate this to c{4,8}, c{6,8}, c{5,7}, and c{5,9} to test edge cases. Invalid cases are skipped. E.g. when the end value becomes less than the start c{4,3} or the start becomes lower than 0 c{-1,3}.

Solves part of this issue:
https://github.com/stryker-mutator/stryker-net/issues/1228

* Add greedy unlimited quantifier regex mutator

In regex c{5,} checks if character c matches 5 to unlimited times consecutively. We can mutate this to c{4,} and c{6,}  to test edge cases. The mutation is skipped when the value becomes lower than zero.

Solves the final part of this issue:
https://github.com/stryker-mutator/stryker-net/issues/1228

Co-authored-by: Pim van Gurp <pim@pimify.nl>
Co-authored-by: Richard Werkman <Richard1158@gmail.com>