Add regex look ahead and behind mutators
In regex there is a way to match something, only if it is preceded or proceded by a group.
For example: (?<=foo)bar will match the word bar in foobar, but won't match bar in minibar.

This mutator flips these regexes to check at the end instead of beginning, and to ensure
that it does NOT start with the group. It flips the check.

Closes #1278