feat(InputFileResolver) : Validate Project File Option Value
ValidateProjectUnderTestNameFilter checks if the filter looks
like a path to a project rather than just a project name (with or without file extension), the exception message will instruct the user in how to properly use the --project-file option.

Unit test added to verify that the exception is thrown when a path is discovered.

This hopefully "fixes" #520 by providing a more helpful exception message.