Fix flaky tests by not running them in parallel
I observed that one of the issues was related to the static `JsonReport._report` field. It would be better to get rid of this static caching, by passing the state around explicitly instead, but for now, fixing the flaky tests is probably still a good idea.

I've had the test running with `while dotnet test --no-build; do :; done` for several minutes now without failure.