Fix finding the project under test
If the test projects are defined in the stryker-config.json file (under stryker-config.test-projects), it becomes impossible to find the project under test unless it is in the current working directory since `FindProjectFile` is used instead of `FindProjectUnderTest`.

It used to work before #1039 where a behavioural change was introduced. @richardwerkman do you remember what was the intention behind that change?