fix(azurefilesharebaselineprovider): removed SetWritingHeaders to fix Signed Version Compatibility
Removed a set of Headers that were beeing appended to http requests when using AzureFileShare as a
provider for Baseline. The headers `x-ms-file-permission`, `x-ms-file-attributes`,
`x-ms-file-creation-time` and `x-ms-file-last-write-time` are now deemed to be optional, and as a
result can be removed. By removing these headers, we are allowing ourselves to be able to use SAS
tokens that have been generated with a Signed Version that is lower than `2019-02-02`.

fix #2114