feat(reporting): add realtime reporting concept
Sends data to the browser, for now this only includes mutation type
including the status of that mutation.

This is a conceptual implementation, this was made to see how easy (or
hard) it would be to implement this (not very hard).

Subsequent commits will improve this solution and increase the data sent
to the browser (complete mutant data is a nice start).